[compiler] Propogate effects stored in object props
Previously, we propogated only effects of functions/methods to the outer
function context.

In this PR, we have extended the propagation of function/method effects
to include those captured in object expressions. This change is
generally a no-op, as the functions defined within the same context as
the object expression already had their effects captured.

This new infra allows us to delete certain effects, such as global
mutation, for object expressions that are passed down as props.

ghstack-source-id: aaa6e87ef953fb25351e84429b8eee9c13ca07ff
Pull Request resolved: https://github.com/facebook/react/pull/30457