[compiler] Delete global mutation effects for objects passed as props
With the tracked list of effects from the previous PR, we delete the
global mutation effects of object expressions that are passed as props.

To make this deletion easier, this PR changes the FunctionEffects to be
a Set.

While using the identity of the effect for comparison may seem less than
ideal, adding extra infrastructure for this purpose doesn't seem
worthwhile for now.

ghstack-source-id: 0ef1609a6e37d01cb20bebbfb7ff1244b01fa031
Pull Request resolved: https://github.com/facebook/react/pull/30458